% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_min_specificity.r
\name{estimate_min_specificity}
\alias{estimate_min_specificity}
\title{Estimate minimum specificity from population testing}
\usage{
estimate_min_specificity(
  positive,
  total,
  cutoff = 0.95,
  samples = 1000,
  min_specificity = 0.997,
  max_specificity = 1,
  prevalence = 0
)
}
\arguments{
\item{positive}{vector of numbers; the number of positive tests in each sample}

\item{total}{vector of numbres; the total number of tests (positive and negative) in each sample}

\item{cutoff}{the cutoff for the estimate, i.e. the confidence level desired}

\item{samples}{number of values of specificity to sample}

\item{min_specificity}{minimum specificity to test}

\item{max_specificity}{maximum specificity to test}

\item{prevalence}{background prevalence; if >0 then it is assumed that
sensitivity is 0, i.e. all positives are still false positives and some of
the negatives are false negatives}
}
\value{
estimate of the lower bound of specificity at
the desired confidence level
}
\description{
Produces an estimate of minimum specificty under the assumption that of the tests identified only negatives as positives
}
\author{
Sebastian Funk
}
